認証トークン取得とAPIリクエスト用のユーティリティ関数を整理し、関数名を変更
