+++ apps/web/src/lib/utils/fetch/fetch-utils.ts
@@ -1,7 +1,9 @@
+import { AUTH_TOKEN_KEY } from '@/constants/auth-token-key'
 import {
   type TemplateResponse,
   templateResponseSchema,
 } from '@ai-job-interview/packages/schemas/common/template-response'
+import { cookies } from 'next/headers'
 
 export function buildApiUrl(path: string): string {
   const apiBaseUrl = process.env.API_BASE_URL || 'http://localhost:4000'
@@ -30,4 +32,18 @@ export async function extractJsonResponse(response: Response) {
   }
 
   return { response, body: validatedBody }
+}
+
+/**
+ * 認証トークンを取得する
+ */
+export async function getAuthToken(): Promise<string> {
+  const cookieStore = await cookies()
+  const token = cookieStore.get(AUTH_TOKEN_KEY)
+
+  if (!token?.value) {
+    throw new Error('認証トークンが見つかりません')
+  }
+
+  return token.value
 }
